Removed old test_git_fat, replaced with new
